refactor: move HardwareException into new errors.py
